# Rewrite selector causing performance issues