--- a/src/components/styled/stat.css
+++ b/src/components/styled/stat.css
@@ -4,7 +4,7 @@
 :where(.stats) {
   @apply divide-x divide-y-0 overflow-x-auto;
 }
-[dir="rtl"] .stats > *:not([hidden]) + *:not([hidden]) {
+[dir="rtl"] .stats > *:not([hidden]) ~ *:not([hidden]) {
   --tw-divide-x-reverse: 1;
 }
 .stat {